+++ IOS/Components/RestaurantRow.swift
@@ -52,6 +52,7 @@ struct RestaurantRow: View {
             priceRange: "$$",
             rating: 4.5,
             distance: 1500,
+            imageURL: "",
             address: Address(id: "a1", street: "Main St", city: "City", district: "District", neighborhood: "Neighborhood"),
             tags: [Tag(id: "t1", name: "Fast Food")],
             promotions: []
