+++ IOS/IOSTests/BusinessServiceTests.swift
@@ -3,7 +3,18 @@ import XCTest
 
 final class BusinessServiceTests: XCTestCase {
     func testGetAllBusinessesSuccess() async throws {
-        let dto = BusinessDTO(id: "1", name: "T", description: nil, priceRange: nil, avgRating: 4.0, address: nil, tags: [], promotions: [])
+        let dto = BusinessDTO(
+            id: "1",
+            name: "T",
+            description: nil,
+            priceRange: nil,
+            avgRating: 4.0,
+            address: nil,
+            tags: [],
+            promotions: [],
+            photos: nil,
+            distance: nil
+        )
         let mock = MockAPIClient()
         mock.result = [dto]
         let service = BusinessService(api: mock)
